fix: - moved `sort-classes` files outside of its folder. - Fixed documentation typo.
