+++ docs/content/rules/sort-classes.mdx
@@ -239,7 +239,7 @@ The `abstract` modifier is incompatible with the `static`, `private` and `decora
 The `abstract` modifier is incompatible with the `static`, `private` and `decorated` modifiers.
 The `declare` modifier is incompatible with the `override` and `decorated` modifiers.
 
-#### Index-signature
+#### Index-signatures
 - Selector: `index-signature`.
 - Modifiers: `readonly`, `static`.
 - Example: `static-readonly-index-signature`.
