+++ index.ts
@@ -9,7 +9,6 @@ import sortSvelteAttributes from './rules/sort-svelte-attributes'
 import sortAstroAttributes from './rules/sort-astro-attributes'
 import sortArrayIncludes from './rules/sort-array-includes'
 import sortVueAttributes from './rules/sort-vue-attributes'
-import sortClasses from './rules/sort-classes/sort-classes'
 import sortNamedImports from './rules/sort-named-imports'
 import sortNamedExports from './rules/sort-named-exports'
 import sortObjectTypes from './rules/sort-object-types'
@@ -17,6 +16,7 @@ import sortSwitchCase from './rules/sort-switch-case'
 import sortUnionTypes from './rules/sort-union-types'
 import sortInterfaces from './rules/sort-interfaces'
 import sortJsxProps from './rules/sort-jsx-props'
+import sortClasses from './rules/sort-classes'
 import sortImports from './rules/sort-imports'
 import sortExports from './rules/sort-exports'
 import sortObjects from './rules/sort-objects'
