+++ rules/sort-classes.ts
@@ -1,23 +1,23 @@
 import type { TSESTree } from '@typescript-eslint/types'
 import type { TSESLint } from '@typescript-eslint/utils'
 
-import type { SortingNode } from '../../typings'
+import type { SortingNode } from '../typings'
 
-import { isPartitionComment } from '../../utils/is-partition-comment'
-import { getCommentBefore } from '../../utils/get-comment-before'
-import { createEslintRule } from '../../utils/create-eslint-rule'
-import { getGroupNumber } from '../../utils/get-group-number'
+import { isPartitionComment } from '../utils/is-partition-comment'
+import { getCommentBefore } from '../utils/get-comment-before'
+import { createEslintRule } from '../utils/create-eslint-rule'
 import { generateOfficialGroups } from './sort-classes-utils'
-import { getSourceCode } from '../../utils/get-source-code'
-import { toSingleLine } from '../../utils/to-single-line'
-import { rangeToDiff } from '../../utils/range-to-diff'
-import { isPositive } from '../../utils/is-positive'
-import { useGroups } from '../../utils/use-groups'
-import { sortNodes } from '../../utils/sort-nodes'
-import { makeFixes } from '../../utils/make-fixes'
-import { complete } from '../../utils/complete'
-import { pairwise } from '../../utils/pairwise'
-import { compare } from '../../utils/compare'
+import { getGroupNumber } from '../utils/get-group-number'
+import { getSourceCode } from '../utils/get-source-code'
+import { toSingleLine } from '../utils/to-single-line'
+import { rangeToDiff } from '../utils/range-to-diff'
+import { isPositive } from '../utils/is-positive'
+import { useGroups } from '../utils/use-groups'
+import { sortNodes } from '../utils/sort-nodes'
+import { makeFixes } from '../utils/make-fixes'
+import { complete } from '../utils/complete'
+import { pairwise } from '../utils/pairwise'
+import { compare } from '../utils/compare'
 
 type MESSAGE_ID = 'unexpectedClassesOrder'
